Backend partially fixed.

+++ es.client/src/main/java/elastest/loganalyzer/es/client/rest/ExecutionRest.java
@@ -3,6 +3,8 @@ package elastest.loganalyzer.es.client.rest;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -13,32 +15,33 @@ import elastest.loganalyzer.es.client.service.ExecutionService;
 import elastest.loganalyzer.es.client.service.LogService;
 
 @RestController
-@RequestMapping("/executions")
+@RequestMapping("/api/executions")
 public class ExecutionRest {
 
-	private final ExecutionService executionService;
-	private final LogService logService;
-
 	@Autowired
-	public ExecutionRest(ExecutionService executionService, LogService logService) {
-		this.executionService = executionService;
-		this.logService = logService;
-	}
+	private ExecutionService executionService;
+	@Autowired
+	private LogService logService;
 
 	@RequestMapping(value = "/project/{project}", method = RequestMethod.GET)
-	public List<Execution> getByProject(@PathVariable String project) {
-		return executionService.findByProjectOrderById(project);
+	public ResponseEntity<List<Execution>> getByProject(@PathVariable String project) {
+		return new ResponseEntity<>(executionService.findByProjectOrderById(project), HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/test/{test}", method = RequestMethod.GET)
-	public Execution getByTest(@PathVariable String test) {
-		return executionService.findByTestId(test);
+	public ResponseEntity<Execution> getByTest(@PathVariable String test) {
+		return new ResponseEntity<>(executionService.findByTestId(test), HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/id/{id}", method = RequestMethod.DELETE)
-	public void deleteById(@PathVariable int id) {
+	public ResponseEntity<Execution> deleteById(@PathVariable int id) {
 		Execution execution = executionService.findOne(id);
-		logService.deleteIterable(logService.findByTestOrderByIdAsc(execution.getTest()));
-		executionService.delete(execution);
+		if (execution != null) {
+			logService.deleteIterable(logService.findByTestOrderByIdAsc(execution.getTest()));
+			executionService.delete(execution);
+			return new ResponseEntity<>(execution, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 }
