+++ es.client/src/main/java/elastest/loganalyzer/es/client/rest/ProjectRest.java
@@ -4,7 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,11 +18,13 @@ import elastest.loganalyzer.es.client.service.LogService;
 import elastest.loganalyzer.es.client.service.ProjectService;
 
 @RestController
-@RequestMapping("/projects")
+@RequestMapping("/api/projects")
 public class ProjectRest {
 
-	private final LogService logService;
-	private final ProjectService projectService;
+	@Autowired
+	private LogService logService;
+	@Autowired
+	private ProjectService projectService;
 
 	@Autowired
 	public ProjectRest(LogService logService, ProjectService projectService) {
@@ -28,32 +32,36 @@ public class ProjectRest {
 		this.projectService = projectService;
 	}
 
-	@RequestMapping(value = "/all", method = RequestMethod.GET)
-	public List<Project> getAll() {
+	@RequestMapping(value = "", method = RequestMethod.GET)
+	public ResponseEntity<List<Project>> getAll() {
 		Iterable<Project> projects = projectService.findAll();
 		List<Project> result = new ArrayList<Project>();
 		for (Project project : projects) {
 			result.add(project);
 		}
-		return result;
+		return new ResponseEntity<>(result, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/name/{name}", method = RequestMethod.GET)
-	public Project getProject(@PathVariable String name) {
-		return projectService.findByName(name);
+	public ResponseEntity<Project> getByName(@PathVariable String name) {
+		return new ResponseEntity<>(projectService.findByName(name), HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public int addLocation(@RequestBody Project project) {
+	public ResponseEntity<Integer> post(@RequestBody Project project) {
 		project.setRecently_deleted(-1);
-		return projectService.save(project);
+		return new ResponseEntity<>(projectService.save(project), HttpStatus.CREATED);
 	}
 
-	@RequestMapping(value = "/remove/id/{id}", method = RequestMethod.DELETE)
-	public Project delete(@PathVariable int id) {
+	@RequestMapping(value = "/id/{id}", method = RequestMethod.DELETE)
+	public ResponseEntity<Project> deleteById(@PathVariable int id) {
 		Project deleted = projectService.findOne(id);
-		logService.deleteIterable(logService.findByProject(deleted.getName()));
-		projectService.delete(deleted);
-		return deleted;
+		if (deleted != null) {
+			logService.deleteIterable(logService.findByProject(deleted.getName()));
+			projectService.delete(deleted);
+			return new ResponseEntity<>(deleted, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 }
